chore(structure): reorganize components and styles into new folders
